ml_training/requirements.txt
mlx>=0.15.0
numpy
Pillow
tqdm
scikit-learn
matplotlib